# jobbing/swagger_remote/swagger.yaml
openapi: 3.0.0
info:
  title: Aprende tu mismo API [remote]
  description: Application created to provide training courses
  termsOfService: https://www.youtube.com/channel/UCz3p_Kz87WePeRMy6x5CX5w
  contact:
    name: Jalisco con empleo
    url: https://bdalimentos.org/
    email: comunicacionbamx@bdalimentos.org
  license:
    name: Apache 2.0
    url: http://apache.com
  version: 1.0.0
servers:
- url: https://www.jaliscoconempleo.org
  description: Production Server
- url: https://127.0.0.1:4010
  description: Development server (uses test data)
paths:
...
  /media/{media_id}:
    get:
      tags:
      - Media
      description: Get media by id
      operationId: get_media_by_id
      parameters:
      - name: media_id
        in: path
        description: Media Id
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses: 
        "200":
          description: successfull response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Media'
        "400":
          description: BAD_REQUEST
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"BAD_REQUEST\"}"
                x-content-type: application/json
        "401":
          description: UNAUTHORIZED
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"Unauthorized\"}"
                x-content-type: application/json
        "404":
          description: Page not found
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"Error 404, page not found.\"}"
                x-content-type: application/json
        "500":
          description: INTERNAL_SERVER_ERROR
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"unexpected error\"}"
                x-content-type: application/json  
      x-openapi-router-controller: jobbing.controllers_remote.media_controller
    delete:
      tags:
      - Media
      description: Delete media by id
      operationId: delete_media_by_id
      parameters:
      - name: media_id
        in: path
        description: Media Id
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses: 
        "200":
          description: successfull response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Media'
        "400":
          description: BAD_REQUEST
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"BAD_REQUEST\"}"
                x-content-type: application/json
        "401":
          description: UNAUTHORIZED
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"Unauthorized\"}"
                x-content-type: application/json
        "404":
          description: Page not found
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"Error 404, page not found.\"}"
                x-content-type: application/json
        "500":
          description: INTERNAL_SERVER_ERROR
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"unexpected error\"}"
                x-content-type: application/json  
      x-openapi-router-controller: jobbing.controllers_remote.media_controller
  /media/{media_id}/content:
    get:
      tags:
      - Media
      description: Get media content by id
      operationId: get_media_content_by_id
      parameters:
      - name: media_id
        in: path
        description: Media Id
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses: 
        "200":
          description: successfull response
          content:
            image/png:
              schema:
                $ref: '#/components/schemas/MediaImageTag'
                format: binary
        "400":
          description: BAD_REQUEST
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"BAD_REQUEST\"}"
                x-content-type: application/json
        "401":
          description: UNAUTHORIZED
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"Unauthorized\"}"
                x-content-type: application/json
        "404":
          description: Page not found
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"Error 404, page not found.\"}"
                x-content-type: application/json
        "500":
          description: INTERNAL_SERVER_ERROR
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"unexpected error\"}"
                x-content-type: application/json  
      x-openapi-router-controller: jobbing.controllers_remote.media_controller
  /media:
    post:
      tags:
      - Media
      description: Post a new media
      operationId: post_media
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Media'
        required: true
      responses:
        "200":
          description: successful request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Media'
        "201":
          description: successful request
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"Execution of update query was successful\"}"
                x-content-type: application/json  
        "204":
          description: successful request
        "400":
          description: BAD_REQUEST
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"BAD_REQUEST\"}"
                x-content-type: application/json
        "401":
          description: UNAUTHORIZED
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"Unauthorized\"}"
                x-content-type: application/json
        "500":
          description: INTERNAL_SERVER_ERROR
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"unexpected error\"}"
                x-content-type: application/json
      x-openapi-router-controller: jobbing.controllers_remote.media_controller
    put:
      tags:
      - Media
      description: Updated a media
      operationId: put_media
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Media'
        required: true
      responses:
        "200":
          description: successful request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Media'
        "201":
          description: successful request
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"Execution of update query was successful\"}"
                x-content-type: application/json  
        "204":
          description: successful request
        "400":
          description: BAD_REQUEST
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"BAD_REQUEST\"}"
                x-content-type: application/json
        "401":
          description: UNAUTHORIZED
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"Unauthorized\"}"
                x-content-type: application/json
        "500":
          description: INTERNAL_SERVER_ERROR
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"unexpected error\"}"
                x-content-type: application/json
      x-openapi-router-controller: jobbing.controllers_remote.media_controller
...
components:
  schemas:
    MediaImageTag:
      type: string
      example: "data:image/jpeg;base64,media_content_base64"
...
    Media:
      required:
      - media_id
      - media_status_id
      - media_data
      type: object
      properties:
        media_id: 
          type: integer
          example: 1
        media_status_id:
          type: integer
          example: 1
        media_data:
          type: string
          example: "0"
        media_link:
          type: string
          example: "http://medialink.com"
        media_title:
          type: string
          example: "user123_profilepicture.png"
        media_description:
          type: string
          example: "User [123] Profile Picture"
        media_size:
          type: number
          example: 123.45
        media_content_upload_date:
          type: string
          example: "2021-12-29T00:00:00Z"
        media_content_updated_date:
          type: string
          example: "2021-12-29T00:00:00Z"
      example:
        media_id: 1
        media_status_id: 1
        media_data: "binfile"
        media_link: "http://medialink.com"
        media_title: "user123_profilepicture.png"
        media_description: "User [123] Profile Picture"
        media_size: 123.45
        media_content_upload_date: "2021-12-29T00:00:00Z"
        media_content_updated_date: "2021-12-29T00:00:00Z"
...